feat(Projects): Icons
update icons
