+++ src/components/Views/Projects/Projects.scss
@@ -86,6 +86,7 @@
           padding: 0.5rem;
           border-right: 1px solid #1e2d3d;
           width: fit-content;
+          white-space: pre-wrap;
         };
 
         .page-mainview {
